Refactor GtfsMetadata into validation package
* This is a better place for it to live, as it directly supports validation
