+++ gtfs-realtime-validator-lib/src/main/java/edu/usf/cutr/gtfsrtvalidator/lib/validation/interfaces/FeedEntityValidator.java
@@ -18,8 +18,8 @@
 package edu.usf.cutr.gtfsrtvalidator.lib.validation.interfaces;
 
 import com.google.transit.realtime.GtfsRealtime;
-import edu.usf.cutr.gtfsrtvalidator.lib.GtfsMetadata;
 import edu.usf.cutr.gtfsrtvalidator.lib.model.helper.ErrorListHelperModel;
+import edu.usf.cutr.gtfsrtvalidator.lib.validation.GtfsMetadata;
 import org.onebusaway.gtfs.impl.GtfsDaoImpl;
 
 import java.util.List;
