+++ gtfs-realtime-validator-webapp/src/main/java/edu/usf/cutr/gtfsrtvalidator/background/BackgroundTask.java
@@ -21,11 +21,11 @@ import com.google.transit.realtime.GtfsRealtime;
 import edu.usf.cutr.gtfsrtvalidator.api.resource.GtfsFeed;
 import edu.usf.cutr.gtfsrtvalidator.db.GTFSDB;
 import edu.usf.cutr.gtfsrtvalidator.helper.DBHelper;
-import edu.usf.cutr.gtfsrtvalidator.lib.GtfsMetadata;
 import edu.usf.cutr.gtfsrtvalidator.lib.model.GtfsRtFeedIterationModel;
 import edu.usf.cutr.gtfsrtvalidator.lib.model.GtfsRtFeedModel;
 import edu.usf.cutr.gtfsrtvalidator.lib.model.helper.ErrorListHelperModel;
 import edu.usf.cutr.gtfsrtvalidator.lib.util.GtfsUtils;
+import edu.usf.cutr.gtfsrtvalidator.lib.validation.GtfsMetadata;
 import edu.usf.cutr.gtfsrtvalidator.lib.validation.interfaces.FeedEntityValidator;
 import edu.usf.cutr.gtfsrtvalidator.lib.validation.rules.*;
 import org.apache.commons.io.IOUtils;
